requirements.txt
competitions-scheduler==0.2.5.post2
dj-database-url==0.5.0
Django==2.1.4
django-bootstrap3==9.1.0
django-bootstrap4==0.0.6
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
pytz==2017.3
whitenoise==4.1.2
django-heroku